Rearranged functions to suitable files
Dependency Functions
