+++ app.js
@@ -46,7 +46,7 @@ app.get('/', function(req,res){
 app.post('/task/add', tq.createTask);
 app.post('/task/delete', tq.deleteTask);
 app.post("/task/update", tq.updateTask)
-app.post("/dependency/add", tq.createDependency)
+app.post("/dependency/add", dq.createDependency)
 app.post("/dependency/update", dq.updateDependency)
 app.post('/dependency/delete', dq.deleteDependency);
